requirements.txt
# Core ML and AI Libraries
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0

# ControlNet and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

# Image Manipulation and Effects
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
scikit-image>=0.21.0

# 3D and Mockup Generation
matplotlib>=3.7.0
plotly>=5.15.0
trimesh>=3.20.0
pyrender>=0.1.45

# Configuration and Utilities
PyYAML>=6.0
click>=8.1.0
tqdm>=4.65.0
rich>=13.0.0
colorama>=0.4.6

# File Handling and I/O
pathlib2>=2.3.7
requests>=2.31.0
aiofiles>=23.1.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: GPU Optimization
xformers>=0.0.20  # For memory optimization
bitsandbytes>=0.41.0  # For quantization

# Optional: Advanced Features
einops>=0.6.0  # For tensor operations
omegaconf>=2.3.0  # For configuration management
hydra-core>=1.3.0  # For experiment management 